refactor: 파일 이름 변경
- 기존의 하나의 함수의 이름이었던 파일들의 이름 변경
