+++ packages/lib/src/index.ts
@@ -1,4 +1,4 @@
-export * from './create-linked-list';
-export * from './create-tree';
 export * from './heap';
+export * from './linked-list';
 export * from './range';
+export * from './tree';
