+++ packages/lib/src/linked-list.ts
