+++ packages/lib/src/tree.ts
